change folder structure and add babel-plugin-module-resolver for import path
