+++ package.json
@@ -30,6 +30,7 @@
     "@babel/core": "^7.20.0",
     "@types/react": "~18.2.45",
     "@types/react-native": "^0.73.0",
+    "babel-plugin-module-resolver": "^5.0.2",
     "typescript": "^5.1.3"
   },
   "private": true
