+++ src/app/(auth)/register.tsx
