+++ src/app/index.tsx
@@ -9,8 +9,8 @@ import {
   View,
 } from 'react-native'
 
-import BackgroundImage from '../assets/images/mainpage.png'
-import ClippedRectangle from '../assets/images/mainpage2.png'
+import BackgroundImage from '@/assets/images/mainpage.png'
+import ClippedRectangle from '@/assets/images/mainpage2.png'
 
 const { width, height } = Dimensions.get('window')
