+++ src/assets/icons/courses.png
