+++ src/components/screens/About.tsx
@@ -1,11 +1,11 @@
 import React from 'react';
 import { Text, View, TouchableOpacity } from 'react-native';
-import { NavigationProp } from '../../types/interfaces';
+import { NavigationProp } from '@/types/interfaces';
 
-const Gallery: React.FC<NavigationProp> = ({ navigate }) => {
+const About: React.FC<NavigationProp> = ({ navigate }) => {
   return (
     <View>
-      <Text>Gallery</Text>
+      <Text>About</Text>
       <TouchableOpacity onPress={() => navigate('Profile')}>
         <Text>Go to Profile</Text>
       </TouchableOpacity>
@@ -13,4 +13,4 @@ const Gallery: React.FC<NavigationProp> = ({ navigate }) => {
   );
 };
 
-export default Gallery;
+export default About;
