+++ src/components/screens/navigationOptions.ts
@@ -1,13 +1,13 @@
-import { ScreenNames } from '../../types/interfaces';
-import AboutIcon from '../../assets/icons/profile-screen/About.svg';
-import CentersIcon from '../../assets/icons/profile-screen/Centers.svg';
-import GalleryIcon from '../../assets/icons/profile-screen/Gallery.svg';
-import CareerIcon from '../../assets/icons/profile-screen/Career.svg';
-import ContactIcon from '../../assets/icons/profile-screen/Contact.svg';
-import TermsIcon from '../../assets/icons/profile-screen/Terms.svg';
-import RefundIcon from '../../assets/icons/profile-screen/Refund.svg';
-import PrivacyIcon from '../../assets/icons/profile-screen/Privacy.svg';
-import LogoutIcon from '../../assets/icons/profile-screen/Logout.svg';
+import { ScreenNames } from '@/types/interfaces';
+import AboutIcon from '@/assets/icons/profile-screen/About.svg';
+import CentersIcon from '@/assets/icons/profile-screen/Centers.svg';
+import GalleryIcon from '@/assets/icons/profile-screen/Gallery.svg';
+import CareerIcon from '@/assets/icons/profile-screen/Career.svg';
+import ContactIcon from '@/assets/icons/profile-screen/Contact.svg';
+import TermsIcon from '@/assets/icons/profile-screen/Terms.svg';
+import RefundIcon from '@/assets/icons/profile-screen/Refund.svg';
+import PrivacyIcon from '@/assets/icons/profile-screen/Privacy.svg';
+import LogoutIcon from '@/assets/icons/profile-screen/Logout.svg';
 
 interface NavigationOption {
   name: ScreenNames;
